3. Spring Security - система регистрации

+++ src/main/java/com/DanielSafonov/Sweater/repos/MessageRepo.java
@@ -6,8 +6,9 @@ import com.DanielSafonov.Sweater.domain.Message;
 
 import java.util.List;
 
+//Доступ к сообщениям
 //Интерфейс для работы с БД
 //Наследует CRUD-функционал
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findAllByTag(String tag); //Метод поиска сообщений по тэгу
 }
